fix(fe/sorting): move sorting function to common
