+++ src/client/web/src/components/layout/rows.tsx
@@ -3,8 +3,8 @@ import { List } from "immutable";
 
 import { BiSortUp } from "@react-icons/all-files/bi/BiSortUp";
 
-
 import { Flexbox } from "./flexbox";
+import { sortRows } from "../../common/utils";
 
 export interface Row {
   elem: React.ReactNode; // element to display
@@ -49,25 +49,7 @@ export class Rows extends React.Component<Props, State, {}> {
     }
     const expectedOrder = !currentOrder;
 
-    const sortedRows = this.props.rows.sort((row1: Row, row2: Row) => {
-      const val1 = row1.sortVals.get(key);
-      const val2 = row2.sortVals.get(key);
-      if (val1 == null || val2 == null) {
-        // elements without the sort key will be moved to the last
-        if (val1 == null && val2 != null) {
-          return 1;
-        } else if (val1 != null && val2 == null) {
-          return -1;
-        }
-        return 0;
-      } else if (val1 < val2) {
-        return expectedOrder ? -1 : 1;
-      } else if (val1 === val2) {
-        return 0;
-      }
-      return expectedOrder ? 1 : -1;
-    });
-
+    const sortedRows = sortRows(this.props.rows, key, expectedOrder);
     const sortedItems = sortedRows.map((row: Row): Object => {
       return row.val;
     });
@@ -104,10 +86,7 @@ export class Rows extends React.Component<Props, State, {}> {
         <div className="margin-b-l">
           <Flexbox
             children={List([
-              <BiSortUp
-                size="3rem"
-                className="black-font margin-r-m"
-              />,
+              <BiSortUp size="3rem" className="black-font margin-r-m" />,
               <span>{sortBtns}</span>,
             ])}
             childrenStyles={List([{ flex: "0 0 auto" }, { flex: "0 0 auto" }])}
